refactor: RankType 패키지 이동
