+++ src/main/java/lotto/config/RankType.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.config;
 
 import java.util.Arrays;
